move identical layer templates to common
