+++ dory/Hardware_targets/PULP/GAP8/HW_description.json
@@ -37,7 +37,7 @@
 	"accelerator frequency": 100000000,
 	"HW specific parameters":{
 		"accelerator core0 stack": 3500,
-		"accelerator core1-7 stack": 3400		
+		"accelerator core1-7 stack": 3400
 	},
     "double_buffering": 1,
     "split_ints": true,
