+++ dory/Hardware_targets/PULP/GAP8/Templates/layer_templates/layer_L2_h_template.h
@@ -1,26 +1,1 @@
-/*
- * layer_template_h.h
- * Alessio Burrello <alessio.burrello@unibo.it>
- *
- * Copyright (C) 2019-2020 University of Bologna
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-#ifndef __${func_name.upper()}_H__
-#define __${func_name.upper()}_H__
-
-void  ${func_name}(
-  void *args
-);
-
-#endif
+../../../Common/Templates/layer_templates/layer_L2_h_template.h
